refactor: Group sessions.
