+++ src/ansys/fluent/core/launcher/container_launcher.py
@@ -68,11 +68,11 @@ from ansys.fluent.core.session import _parse_server_info_file
 from ansys.fluent.core.utils.fluent_version import FluentVersion
 
 if TYPE_CHECKING:
-    from ansys.fluent.core.session_meshing import Meshing
-    from ansys.fluent.core.session_pure_meshing import PureMeshing
-    from ansys.fluent.core.session_solver import Solver
-    from ansys.fluent.core.session_solver_aero import SolverAero
-    from ansys.fluent.core.session_solver_icing import SolverIcing
+    from ansys.fluent.core.session.session_meshing import Meshing
+    from ansys.fluent.core.session.session_pure_meshing import PureMeshing
+    from ansys.fluent.core.session.session_solver import Solver
+    from ansys.fluent.core.session.session_solver_aero import SolverAero
+    from ansys.fluent.core.session.session_solver_icing import SolverIcing
 
 
 class ContainerArgsWithoutDryRunMode(LauncherArgsBase, TypedDict, total=False):
