+++ src/ansys/fluent/core/launcher/launcher.py
@@ -67,11 +67,11 @@ from ansys.fluent.core.launcher.slurm_launcher import SlurmFuture, SlurmLauncher
 from ansys.fluent.core.launcher.standalone_launcher import StandaloneLauncher
 import ansys.fluent.core.launcher.watchdog as watchdog
 from ansys.fluent.core.module_config import config
-from ansys.fluent.core.session_meshing import Meshing
-from ansys.fluent.core.session_pure_meshing import PureMeshing
-from ansys.fluent.core.session_solver import Solver
-from ansys.fluent.core.session_solver_aero import SolverAero
-from ansys.fluent.core.session_solver_icing import SolverIcing
+from ansys.fluent.core.session.session_meshing import Meshing
+from ansys.fluent.core.session.session_pure_meshing import PureMeshing
+from ansys.fluent.core.session.session_solver import Solver
+from ansys.fluent.core.session.session_solver_aero import SolverAero
+from ansys.fluent.core.session.session_solver_icing import SolverIcing
 from ansys.fluent.core.utils.deprecate import deprecate_arguments
 from ansys.fluent.core.utils.fluent_version import FluentVersion
