+++ src/ansys/fluent/core/launcher/pim_launcher.py
@@ -57,11 +57,11 @@ from ansys.fluent.core.utils.fluent_version import FluentVersion
 import ansys.platform.instancemanagement as pypim
 
 if TYPE_CHECKING:
-    from ansys.fluent.core.session_meshing import Meshing
-    from ansys.fluent.core.session_pure_meshing import PureMeshing
-    from ansys.fluent.core.session_solver import Solver
-    from ansys.fluent.core.session_solver_aero import SolverAero
-    from ansys.fluent.core.session_solver_icing import SolverIcing
+    from ansys.fluent.core.session.session_meshing import Meshing
+    from ansys.fluent.core.session.session_pure_meshing import PureMeshing
+    from ansys.fluent.core.session.session_solver import Solver
+    from ansys.fluent.core.session.session_solver_aero import SolverAero
+    from ansys.fluent.core.session.session_solver_icing import SolverIcing
 
 
 class PIMArgsWithoutMode(
