+++ src/ansys/fluent/core/session/session_utilities.py
@@ -27,13 +27,6 @@ from typing import TYPE_CHECKING, Any, Literal, overload
 
 from typing_extensions import Unpack, override
 
-from ansys.fluent.core import (
-    session_meshing,
-    session_pure_meshing,
-    session_solver,
-    session_solver_aero,
-    session_solver_icing,
-)
 from ansys.fluent.core.launcher.container_launcher import (
     ContainerArgsWithoutDryRunMode,
     DockerLauncher,
@@ -50,7 +43,14 @@ from ansys.fluent.core.launcher.standalone_launcher import (
     StandaloneArgsWithoutDryRunMode,
     StandaloneLauncher,
 )
-from ansys.fluent.core.session import BaseSession
+from ansys.fluent.core.session import (
+    BaseSession,
+    session_meshing,
+    session_pure_meshing,
+    session_solver,
+    session_solver_aero,
+    session_solver_icing,
+)
 
 __all__ = (
     "Meshing",
