+++ tests/test_launcher_remote.py
@@ -47,8 +47,8 @@ from ansys.fluent.core.fluent_connection import (
 )
 from ansys.fluent.core.launcher import launcher
 from ansys.fluent.core.session import BaseSession
-from ansys.fluent.core.session_pure_meshing import PureMeshing
-from ansys.fluent.core.session_solver import Solver
+from ansys.fluent.core.session.session_pure_meshing import PureMeshing
+from ansys.fluent.core.session.session_solver import Solver
 from ansys.fluent.core.utils.file_transfer_service import PimFileTransferService
 import ansys.fluent.core.utils.fluent_version as docker_image_version
 from ansys.fluent.core.utils.fluent_version import FluentVersion
